Extract play length calculation to helper class
