+++ osu.Server.Queues.ScoreStatisticsProcessor/Processors/PlayTimeProcessor.cs
@@ -1,13 +1,10 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Linq;
 using JetBrains.Annotations;
 using MySqlConnector;
 using osu.Game.Online.API.Requests.Responses;
-using osu.Game.Rulesets;
-using osu.Game.Rulesets.Mods;
+using osu.Server.Queues.ScoreStatisticsProcessor.Helpers;
 using osu.Server.Queues.ScoreStatisticsProcessor.Models;
 
 namespace osu.Server.Queues.ScoreStatisticsProcessor.Processors
@@ -25,37 +22,16 @@ namespace osu.Server.Queues.ScoreStatisticsProcessor.Processors
         public void RevertFromUserStats(SoloScoreInfo score, UserStats userStats, int previousVersion, MySqlConnection conn, MySqlTransaction transaction)
         {
             if (previousVersion >= 6)
-                userStats.total_seconds_played -= getPlayLength(score);
+                userStats.total_seconds_played -= PlayValidityHelper.GetPlayLength(score);
         }
 
         public void ApplyToUserStats(SoloScoreInfo score, UserStats userStats, MySqlConnection conn, MySqlTransaction transaction)
         {
-            userStats.total_seconds_played += getPlayLength(score);
+            userStats.total_seconds_played += PlayValidityHelper.GetPlayLength(score);
         }
 
         public void ApplyGlobal(SoloScoreInfo score, MySqlConnection conn)
         {
-        }
-
-        private static int getPlayLength(SoloScoreInfo score)
-        {
-            // to ensure sanity, first get the maximum time feasible from the beatmap's length
-            double totalLengthSeconds = score.Beatmap!.Length;
-
-            Ruleset ruleset = ScoreStatisticsQueueProcessor.AVAILABLE_RULESETS.Single(r => r.RulesetInfo.OnlineID == score.RulesetID);
-
-            var rateAdjustMods = score.Mods.Select(m => m.ToMod(ruleset)).OfType<ModRateAdjust>().ToArray();
-
-            foreach (var mod in rateAdjustMods)
-                totalLengthSeconds /= mod.SpeedChange.Value;
-
-            if (score.StartedAt == null)
-                return (int)totalLengthSeconds;
-
-            // TODO: better handle failed plays once we have incoming data.
-
-            TimeSpan realTimePassed = score.EndedAt - score.StartedAt.Value;
-            return (int)Math.Min(totalLengthSeconds, realTimePassed.TotalSeconds);
         }
     }
 }
